consensus, node/cn: move Protocol to node/cn, guard consensus msg dispatch
